+++ src/feature/post/index.ts
@@ -1,0 +1,2 @@
+export { useEditPost } from './model/hooks/useEditPost'
+export { EditPostForm } from './ui/EditPostForm/EditPostForm'
