--- src/feature/post/ui/EditPost/EditPost.module.scss
@@ -1,36 +1,0 @@
-.block {
-  position: relative;
-
-  width: 481px;
-  height: 500px;
-  padding: 10px 24px;
-
-  background-color: var(--dark-300);
-}
-
-.editPostArea {
-  width: 100%;
-}
-
-.editPostArea textarea {
-  height: 120px;
-}
-
-.valueLength {
-  padding: 5px 0;
-  color: var(--light-900);
-  text-decoration: none;
-}
-
-.blockArea {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  width: 100%;
-}
-
-.btn {
-  position: absolute;
-  right: 24px;
-  bottom: 24px;
-}
