+++ src/shared/ui/TextArea/TextArea.module.scss
@@ -1,6 +1,7 @@
 .container {
   display: flex;
   flex-direction: column;
+  width: 100%;
 }
 
 .label {
